+++ Firestore/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v8.9.0
 - [added] Added support for Firebase AppCheck.
 - [fixed] Fix a crash if `[FIRTransaction getDocument]` was called after
   `[FIRFirestore terminateWithCompletion]` (#8760).